Moves SMF\Event to SMF\Calendar\Event
