+++ Sources/Actions/Post2.php
@@ -20,11 +20,11 @@ use SMF\BBCodeParser;
 use SMF\Board;
 use SMF\BrowserDetector;
 use SMF\Cache\CacheApi;
+use SMF\Calendar\Event;
 use SMF\Config;
 use SMF\Db\DatabaseApi as Db;
 use SMF\Draft;
 use SMF\ErrorHandler;
-use SMF\Event;
 use SMF\IntegrationHook;
 use SMF\Lang;
 use SMF\Logging;
